+++ modules/Students/AssignOtherInfo.php
@@ -50,7 +50,24 @@ if ( $_REQUEST['modfunc'] === 'save'
 			$_REQUEST['values']['ENROLLMENT_CODE']
 		);
 
-		foreach ( (array) $_REQUEST['student'] as $student_id )
+		$student_ids = $_REQUEST['student'];
+
+		if ( User( 'SCHOOLS' ) )
+		{
+			// @since 12.9 Security fix #376 IDOR: Unauthorized Cross-School Data Assignment
+			$students_list = implode( ',', array_map( 'intval', $_REQUEST['student'] ) );
+
+			$students_RET = DBGet( "SELECT DISTINCT STUDENT_ID
+				FROM student_enrollment
+				WHERE SYEAR='" . UserSyear() . "'
+				AND SCHOOL_ID IN(" . mb_substr( str_replace( ',', "','", User( 'SCHOOLS' ) ), 2, -2 ) . ")
+				AND STUDENT_ID IN(" . $students_list . ")",
+				[], [ 'STUDENT_ID' ] );
+
+			$student_ids = empty( $students_RET ) ? [] : array_keys( $students_RET );
+		}
+
+		foreach ( (array) $student_ids as $student_id )
 		{
 			$update_enrollment_col_final = $update_enrollment_col;
 
@@ -138,7 +155,7 @@ if ( $_REQUEST['modfunc'] === 'save'
 			'students',
 			$update_student_col,
 			// @since 12.7 SQL allow array in $where_columns: WHERE COLUMN IN(val1,val2)
-			[ 'STUDENT_ID' => array_map( 'intval', $_REQUEST['student'] ) ]
+			[ 'STUDENT_ID' => array_map( 'intval', $student_ids ) ]
 		);
 
 		if ( ! $update_enrollment_col