+++ CHANGES.md
@@ -26,7 +26,8 @@ Changes in 12.9
 - Remove `&modfunc=` from link, never actually used in Assignments.php, MassCreateAssignments.php, Courses.php & MarkingPeriods.php
 - Skip CSP violation triggered by domain: overbridgenet.com, .goguardian, tl.ytlogs.ru, mainf.global-cache.online & wallet.binance.com in SaveReport.php
 - Remove duplicate GET params from form URL (start & end date, others) in Percent.php, CategoryBreakdown.php, CategoryBreakdownTime.php, StudentFieldBreakdown.php, TransactionsReport.php, AddDrop.php, Calendar.php, StudentBreakdown.php & AddDropBreakdownTime.php
-- Security fix #375 Authenticated Time-based SQL Injection (non-student roles) in StudentSumary.php
+- Security fix #375 Authenticated Time-based SQL Injection (non-student roles) in StudentSumary.php, thanks to @Cresp0
+- Security fix #376 IDOR: Unauthorized Cross-School Data Assignment in AddAbsences.php, AddActivity.php, Requests.php, MassAssignFees.php, MassAssignPayments.php, AssignOtherInfo.php & AddStudents.php, thanks to @Cresp0
 
 Changes in 12.8.1
 -----------------